server/requirements.txt
fastapi
uvicorn
python-dotenv
groq
reportlab
pydantic
python-multipart
